Added Global Variable
