+++ opta/special_formatter.py
@@ -2,7 +2,7 @@ import logging
 import string
 from typing import Any
 
-from opta.constants import InterpolationErrors
+from opta.utils.globals import Interpolation
 
 logger = logging.getLogger("opta")
 
@@ -16,7 +16,7 @@ class PartialFormatter(string.Formatter):
         try:
             val = super().get_field(field_name, args, kwargs)
         except (KeyError, AttributeError):
-            InterpolationErrors.add(field_name)
+            Interpolation.add(field_name)
             logger.debug(
                 f"Did not find field {field_name} when trying to format. Odds are "
                 "this is just setting it to an empty default and you can ignore this"
